divided large class into smaller classes; changed the way how env variables are imported
